payload: move manager to a separate file
